[comments] Add delete and update comments methods

+++ src/services/tickets/tickets.service.js
@@ -12,7 +12,9 @@ module.exports = {
     create,
     update,
     delete: _delete,
-    createComment
+    createComment,
+    updateComment,
+    deleteComment
 }
 
 async function getAll(query) {
@@ -35,7 +37,7 @@ async function getById(ticket_id) {
     if (!ObjectId.isValid(ticket_id)) {
         throw 'Provided ticket_id is invalid'
     }
-    
+
     const ticket = await Ticket.findOne(new ObjectId(ticket_id))
 
     if (!ticket) {
@@ -45,13 +47,15 @@ async function getById(ticket_id) {
     return ticket
 }
 
-async function create(ticketParam) {
-    const ticket = new Ticket(ticketParam)
-    
+async function create(ticketParam, user) {
+    var ticket = new Ticket(ticketParam)
+    ticket.creator_id = user.uidNumber
+    ticket.creator_display_name = user.cn
+
     return await ticket.save()
 }
 
-async function update(ticket_id, ticketParam) {
+async function update(ticket_id, ticketParam, user_id) {
     if (!ObjectId.isValid(ticket_id)) {
         throw 'Provided ticket_id is invalid'
     }
@@ -62,46 +66,96 @@ async function update(ticket_id, ticketParam) {
 
     Object.assign(existingTicket, ticketParam)
 
+    existingTicket.updated_date = Date.now
+
     return await existingTicket.save()
 }
 
-async function _delete(ticket_id) {
+async function _delete(ticket_id, user_id) {
     if (!ObjectId.isValid(ticket_id)) {
         throw 'Ticket_id is invalid'
+    }
+
+    const existingTicket = await Ticket.findOne(new ObjectId(ticket_id))
+
+    if (existingTicket.creator_id != user_id) {
+        throw 'You do not have permissions to delete this ticket'
     }
 
     return await Ticket.deleteOne({ '_id': new ObjectId(ticket_id) })
 }
 
-async function createComment(ticket_id, commentParam) {
+async function createComment(ticket_id, commentParam, user) {
     if (!ObjectId.isValid(ticket_id)) {
         throw 'Provided ticket_id is invalid'
     }
 
     const ticket = await Ticket.findOne(new ObjectId(ticket_id))
 
-    if(!ticket){
+    if (!ticket) {
         throw 'Ticket with given ID does not exist'
     }
 
-    ticket.comments.push(ticket.comments.create(commentParam))
-    
+    var comment = ticket.comments.create(commentParam)
+    comment.author_id = user.uidNumber
+    comment.author_display_name = user.cn
+
+    ticket.comments.push(comment)
+
     return await ticket.save()
 }
 
-async function deleteComment(comment_id) {
+async function updateComment(comment_id, appParam, user_id) {
     if (!ObjectId.isValid(comment_id)) {
         throw 'Provided comment_id is invalid'
     }
 
-    const comment = await Ticket.findOne({'comments._id': new ObjectId(comment_id)})
+    const ticket = await Ticket.findOne({ 'comments._id': new ObjectId(comment_id) })
 
-    if(!comment){
+    if (!ticket) {
         throw 'Comment with given ID does not exist'
     }
 
-    return Ticket.findOne({'comments._id': new ObjectId(app_id)}, function(err, result){
-        result.comments.id(comment_id).remove()
-        result.save()
-    })
+    var set = {}
+    for (var field in appParam) {
+        set['comments.$.' + field] = appParam[field]
+    }
+    set['comments.$.updated_at'] = new Date(Date.now())
+
+    const comment = ticket.comments.id(comment_id)
+
+    if (comment.author_id != user_id) {
+        throw 'You do not have permissions to edit this comment'
+    }
+
+    const updatedRecord = await Ticket.findOneAndUpdate({ 'comments._id': new ObjectId(comment_id) }, { $set: set }, { new: true })
+    if (!updatedRecord) {
+        throw 'Comment with given ID does not exist'
+    }
+
+    updatedRecord.save()
+
+    return updatedRecord
+}
+
+async function deleteComment(comment_id, user_id) {
+    if (!ObjectId.isValid(comment_id)) {
+        throw 'Provided comment_id is invalid'
+    }
+
+    const ticket = await Ticket.findOne({ 'comments._id': new ObjectId(comment_id) })
+
+    if (!ticket) {
+        throw 'Comment with given ID does not exist'
+    }
+
+    const comment = ticket.comments.id(comment_id)
+
+    if (comment.author_id != user_id) {
+        throw 'You do not have permissions to delete this comment'
+    }
+
+    await comment.remove()
+
+    return ticket.save()
 }
